Add JSONSchema to Utils for wider usage
